fix support for exposed variadic methods in the job mapper
the essential problem is that some functions (particularly those decorated with
inlineCallbacks) were inheriting a variadic (*args, **kwargs) call signature
from the wrapper function defined in the decorator.  This meant that inspect
could not figure out what was required or not required, and in this case it
assumed there were no required or keyword arguments;  so the underlying functions
would be called iwth no arguments and they wouldn't work at all.

Now in such a situation, the job's kwargs (which is an amalagam of various rows
in a few tables, but mostly the content_*account and the associated
content_account table) are correctly mapped using the reverse mapping for the
service type and are *all* sent to the function as kwargs.  This is actually
not dangerous, but it does mean that functions using inlineCallbacks should
probably accept **kwargs or they will risk being called with excess args.